requirements.txt
python-telegram-bot==20.3
sqlalchemy==2.0.23
requests==2.31.0
matplotlib==3.5.3
pandas==2.0.3
python-dotenv==1.0.0
ydb==3.23.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
numpy==1.26.4